make atom number box like symbol box wth extra characters
this makes atom box display ">" as the last digit when the value is larger than the object's width. This is the same behavior as the symbol box.

Before, it just used to collapse to "+" or "-", or display truncated values without indication. I think this is more useful and coherent to the behavior of symbol box.

But I can see how there was a clear intent to implement the behavior I'm proposing to change, so maybe some discussion is needed as to why the last coded behavior was considered.

Next is making [nbx] behave the same way